# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A simple REST API for Pokemon data",
        "title": "Pokemon API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/pokemon/random": {
            "get": {
                "description": "Fetches a random Pokemon with its data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pokemon"
                ],
                "summary": "Get a random Pokemon",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Pokemon"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Pokemon": {
            "type": "object",
            "properties": {
                "abilities": {
                    "description": "Abilities of the Pokemon",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"static\"]"
                    ]
                },
                "height": {
                    "description": "Height in decimeters",
                    "type": "integer",
                    "example": 4
                },
                "id": {
                    "description": "The unique identifier for the Pokemon",
                    "type": "integer",
                    "example": 25
                },
                "name": {
                    "description": "The name of the Pokemon",
                    "type": "string",
                    "example": "pikachu"
                },
                "sprites": {
                    "description": "Visual representations of the Pokemon",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Sprites"
                        }
                    ]
                },
                "types": {
                    "description": "Types of the Pokemon",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"electric\"]"
                    ]
                },
                "weight": {
                    "description": "Weight in hectograms",
                    "type": "integer",
                    "example": 60
                }
            }
        },
        "models.Sprites": {
            "type": "object",
            "properties": {
                "back_default": {
                    "description": "URL to the back default sprite",
                    "type": "string",
                    "example": "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/back/25.png"
                },
                "front_default": {
                    "description": "URL to the front default sprite",
                    "type": "string",
                    "example": "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/25.png"
                }
            }
        }
    }
}